% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashToPort.R
\name{hashToPort}
\alias{hashToPort}
\title{Hash a value to a valid port}
\usage{
hashToPort(value)
}
\arguments{
\item{value}{Character or numeric (quasi-integer) vector of length \code{1}.}
}
\value{
Integer.
}
\description{
Hash a value to a valid port
}
\details{
The implemented hash function is pure but not perfect/
collision free. This is just a R binding to the fabolous \code{hash-to-port}
npm module bundled by \code{mafintosh} based on the marvelous 
\code{hash-index} npm module written by \code{watson}.
}
\seealso{
\url{https://github.com/mafintosh/hash-to-port} 
\url{https://github.com/watson/hash-index}
}
